StatisticDao setAll() refactored, StatisticsTable double changed to BigDecimal

+++ src/main/java/ui/StatisticsTable.java
@@ -5,6 +5,7 @@ import model.Category;
 import model.CategoryStatistic;
 
 import java.awt.*;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,11 +14,11 @@ public class StatisticsTable extends AbstractEntityTableModel<CategoryStatistic>
     private static final List<Column<?, CategoryStatistic>> COLUMNS = List.of(
             Column.readOnly("Category", String.class, CategoryStatistic::getCategoryName),
             Column.readOnly("Transactions", Integer.class, CategoryStatistic::getTransactionsNumber),
-            Column.readOnly("Income", Double.class, CategoryStatistic::getIncome),
-            Column.readOnly("Expenses", Double.class, CategoryStatistic::getExpenses),
-            Column.readOnly("% of total income", Double.class, CategoryStatistic::getPercentageInc),
-            Column.readOnly("% of total spending", Double.class, CategoryStatistic::getPercentageSpend),
-            Column.readOnly("Sum", Double.class, CategoryStatistic::getSum)
+            Column.readOnly("Income", BigDecimal.class, CategoryStatistic::getIncome),
+            Column.readOnly("Expenses", BigDecimal.class, CategoryStatistic::getExpenses),
+            Column.readOnly("% of total income", BigDecimal.class, CategoryStatistic::getPercentageInc),
+            Column.readOnly("% of total spending", BigDecimal.class, CategoryStatistic::getPercentageSpend),
+            Column.readOnly("Sum", BigDecimal.class, CategoryStatistic::getSum)
     );
 
     private final List<CategoryStatistic> statistics;
